+++ Content.Server/Radio/EntitySystems/RadioDeviceSystem.cs
@@ -1,5 +1,6 @@
 using Content.Server.Chat.Systems;
 using Content.Server.Interaction;
+using Content.Server.Language;
 using Content.Server.Popups;
 using Content.Server.Power.Components;
 using Content.Server.Power.EntitySystems;
