+++ Content.Server/Radio/EntitySystems/RadioSystem.cs
@@ -1,5 +1,6 @@
 using Content.Server.Administration.Logs;
 using Content.Server.Chat.Systems;
+using Content.Server.Language;
 using Content.Server.Power.Components;
 using Content.Server.Radio.Components;
 using Content.Server.Speech;
